Revisit TCP & UDP

+++ modules/tcp.js
@@ -6,31 +6,43 @@ Engine does not send any data using TCP protocol; therefore, this is a one way c
 */
 
 const net = require('net');
+const EventEmitter = require('events');
 
 const tcp_client = new net.Socket();
 
 module.exports = {
 	tcp_connected: false,
+	emitter: new EventEmitter(),
 	connectTCP: function(port, ip) {
 		/*
 		Connects to the server.
 		*/
+
+		// If TCP is running, then skip.
+		if(module.exports.tcp_connected === true) {
+			return;
+		}
+
 		console.log('[TCP] Connecting to '+ip+":"+port+"...");
-		tcp_client.connect(port, ip, function() {
-			console.log('[TCP] Waiting for establishment...');
-		});
+		tcp_client.connect(port, ip);
 	},
 	sendTCP: function(data) {
 		/*
 		Sends a message to the server.
 		*/
+
+		console.log('[TCP] Sent message.');
 		tcp_client.write(data);
-		console.log('[TCP] Sent message ('+data+').');
 	},
 	destroyTCP: function() {
 		/*
 		Destroys the connection.
 		*/
+
+		if(module.exports.tcp_connected === false) {
+			return;
+		}
+
 		tcp_client.destroy();
 	}
 };
@@ -39,6 +51,7 @@ tcp_client.on('data', function(data) {
 	/*
 	Emitted when TCP client receives data from the server.
 	*/
+
 	console.log("[TCP] Received data.");
 	console.log("[TCP] Receiving data is unsupported in RESFET, please make sure that the server is in the right version. Refer to GitHub if necessary.");
 });
@@ -47,14 +60,26 @@ tcp_client.on('close', function(data) {
 	/*
 	Emitted when TCP client is disconnected.
 	*/
-	console.log('[TCP] Connection closed. ('+data+').');
-  	module.exports.tcp_connected = false;
+	console.log('[TCP] Connection closed.');
+	module.exports.tcp_connected = false;
+
+	// Emit TCP status.
+	module.exports.emitter.emit("status", module.exports.tcp_connected);
 });
 
 tcp_client.on('connect', function() {
 	/*
 	Emitted when TCP client connects to the server.
 	*/
-	console.log('[TCP] Connection established.');
-  	module.exports.tcp_connected = true;
+	console.log(`[TCP] Connection established.`);
+	module.exports.tcp_connected = true;
+
+	// Emit TCP status.
+	module.exports.emitter.emit("status", module.exports.tcp_connected);
+});
+
+tcp_client.on('error', function(err) {
+	if(err.code == 'ECONNREFUSED') {
+		console.log(`[TCP] Server could not be reached on ${err.address}:${err.port}.`);
+	}
 });
